requirements.txt
faxplus-api
Flask==0.12.2
requests==2.18.4

